pyproject.toml
[tool.poetry]
name = "genima"
version = "0.1.0"
description = "Generative Image as Action Models"
authors = ["Mohit Shridhar <mohit.shridhar@dyson.com>", "Yat Long (Richie), Lo <richie.lo@dyson.com>", "Stephen James <stephen.james@dyson.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.10.12"
hydra-core = "^1.3.2"
numpy = ">=1.26.4,<2.0.0"
pillow = "10.2.0"
pyrender = "0.1.45"
pyquaternion = "^0.9.9"
scipy = "^1.13.1"
transformers = "4.38.0"
torch = { version = "2.0.1", source = "pytorch" }
torchvision = { version = "0.15.2", source = "pytorch" }
torchaudio = { version = "2.0.2", source = "pytorch" }
cffi = "1.14.2"
matplotlib = "^3.9.0"
absl-py = "1.4.0"
diffusers = "0.29.0"
datasets = "2.20.0"
xformers = "0.0.22"
wandb = "0.14.2"
tqdm = "^4.66.4"
graphql-core = "2.0"
accelerate = "0.25.0"
tiger = {git = "ssh://github.com/stepjam/tiger.git"}
clip = {git = "https://github.com/openai/CLIP.git"}
colosseum = {git = "https://github.com/robot-colosseum/robot-colosseum.git", rev = "9141099aedb7bd31c3b1d090957f73969b4de4cf"}

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-exec-plugin.commands]
rlbench = "./scripts/install_rlbench.sh"
download_textures = "./scripts/download_mil_textures.sh"
